moved google services/firebase to nonfree/foss sourceSets
